+++ library/core/src/main/java/com/google/android/exoplayer2/text/webvtt/WebvttSubtitle.java
@@ -15,7 +15,6 @@
  */
 package com.google.android.exoplayer2.text.webvtt;
 
-import android.text.SpannableStringBuilder;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.text.Cue;
 import com.google.android.exoplayer2.text.Subtitle;
@@ -23,6 +22,7 @@ import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Util;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -30,17 +30,16 @@ import java.util.List;
  */
 /* package */ final class WebvttSubtitle implements Subtitle {
 
-  private final List<Cue> cues;
+  private final List<WebvttCueInfo> cueInfos;
   private final long[] cueTimesUs;
   private final long[] sortedCueTimesUs;
 
   /** Constructs a new WebvttSubtitle from a list of {@link WebvttCueInfo}s. */
   public WebvttSubtitle(List<WebvttCueInfo> cueInfos) {
-    this.cues = new ArrayList<>(cueInfos.size());
+    this.cueInfos = Collections.unmodifiableList(new ArrayList<>(cueInfos));
     cueTimesUs = new long[2 * cueInfos.size()];
     for (int cueIndex = 0; cueIndex < cueInfos.size(); cueIndex++) {
       WebvttCueInfo cueInfo = cueInfos.get(cueIndex);
-      this.cues.add(cueInfo.cue);
       int arrayIndex = cueIndex * 2;
       cueTimesUs[arrayIndex] = cueInfo.startTimeUs;
       cueTimesUs[arrayIndex + 1] = cueInfo.endTimeUs;
@@ -69,53 +68,34 @@ import java.util.List;
 
   @Override
   public List<Cue> getCues(long timeUs) {
-    List<Cue> list = new ArrayList<>();
-    Cue firstNormalCue = null;
-    SpannableStringBuilder normalCueTextBuilder = null;
-
-    for (int i = 0; i < cues.size(); i++) {
+    List<Cue> currentCues = new ArrayList<>();
+    List<WebvttCueInfo> cuesWithUnsetLine = new ArrayList<>();
+    for (int i = 0; i < cueInfos.size(); i++) {
       if ((cueTimesUs[i * 2] <= timeUs) && (timeUs < cueTimesUs[i * 2 + 1])) {
-        Cue cue = cues.get(i);
-        // TODO(ibaker): Replace this with a closer implementation of the WebVTT spec (keeping
-        // individual cues, but tweaking their `line` value):
-        // https://www.w3.org/TR/webvtt1/#cue-computed-line
-        if (isNormal(cue)) {
-          // We want to merge all of the normal cues into a single cue to ensure they are drawn
-          // correctly (i.e. don't overlap) and to emulate roll-up, but only if there are multiple
-          // normal cues, otherwise we can just append the single normal cue.
-          if (firstNormalCue == null) {
-            firstNormalCue = cue;
-          } else if (normalCueTextBuilder == null) {
-            normalCueTextBuilder = new SpannableStringBuilder();
-            normalCueTextBuilder
-                .append(Assertions.checkNotNull(firstNormalCue.text))
-                .append("\n")
-                .append(Assertions.checkNotNull(cue.text));
-          } else {
-            normalCueTextBuilder.append("\n").append(Assertions.checkNotNull(cue.text));
-          }
+        WebvttCueInfo cueInfo = cueInfos.get(i);
+        if (cueInfo.cue.line == Cue.DIMEN_UNSET) {
+          cuesWithUnsetLine.add(cueInfo);
         } else {
-          list.add(cue);
+          currentCues.add(cueInfo.cue);
         }
       }
     }
-    if (normalCueTextBuilder != null) {
-      // There were multiple normal cues, so create a new cue with all of the text.
-      list.add(WebvttCueParser.newCueForText(normalCueTextBuilder));
-    } else if (firstNormalCue != null) {
-      // There was only a single normal cue, so just add it to the list.
-      list.add(firstNormalCue);
+    // Steps 4 - 10 of https://www.w3.org/TR/webvtt1/#cue-computed-line
+    // (steps 1 - 3 are handled by WebvttCueParser#computeLine(float, int))
+    Collections.sort(cuesWithUnsetLine, (c1, c2) -> Long.compare(c1.startTimeUs, c2.startTimeUs));
+    for (int i = 0; i < cuesWithUnsetLine.size(); i++) {
+      Cue cue = cuesWithUnsetLine.get(i).cue;
+      currentCues.add(
+          cue.buildUpon()
+              .setLine((float) (-1 - i), Cue.LINE_TYPE_NUMBER)
+              // WebVTT doesn't use 'line alignment' (i.e. Cue#lineAnchor) when computing position
+              // with snap-to-lines=true (i.e. Cue#LINE_TYPE_NUMBER) but Cue does use lineAnchor
+              // when describing how numeric cues should be displayed. So we have to manually set
+              // lineAnchor=ANCHOR_TYPE_END to avoid the bottom line of cues being off the screen.
+              // https://www.w3.org/TR/webvtt1/#processing-cue-settings
+              .setLineAnchor(Cue.ANCHOR_TYPE_END)
+              .build());
     }
-    return list;
-  }
-
-  /**
-   * Returns whether or not this cue should be placed in the default position and rolled-up with the
-   * other "normal" cues.
-   *
-   * @return Whether this cue should be placed in the default position.
-   */
-  private static boolean isNormal(Cue cue) {
-    return (cue.line == Cue.DIMEN_UNSET && cue.position == WebvttCueParser.DEFAULT_POSITION);
+    return currentCues;
   }
 }